Refactor project structure and add Gradle build system
