+++ .gitignore
@@ -4,7 +4,8 @@ out/
 *.iml
 !**/src/main/**/out/
 !**/src/test/**/out/
-
+build/
+.gradle/
 
 ### Eclipse ###
 .apt_generated
