+++ src/main/java/Main.java
@@ -1,4 +1,4 @@
-package main;
+import view.GamePanel;
 
 import javax.swing.*;
 
@@ -12,7 +12,6 @@ public class Main {
         window.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         window.setResizable(false);
         window.setTitle("Battle City");
-
 
         GamePanel gamePanel = new GamePanel();
         window.add(gamePanel);
