+++ src/main/java/controller/KeyHandler.java
@@ -1,4 +1,4 @@
-package main;
+package controller;
 
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
