+++ src/main/java/controller/MouseHandler.java
@@ -1,11 +1,10 @@
-package main;
+package controller;
 
 import java.awt.*;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 import java.awt.event.MouseWheelEvent;
 import java.awt.event.MouseWheelListener;
-import java.util.Optional;
 
 public class MouseHandler implements MouseListener, MouseWheelListener {
