+++ src/main/java/tile/TileManager.java
@@ -1,6 +1,8 @@
 package tile;
 
-import main.GamePanel;
+
+
+import view.GamePanel;
 
 import javax.imageio.ImageIO;
 import java.awt.*;
