+++ src/main/resources/buildings/house1.png
