+++ src/main/resources/map/map00.txt
